% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{as.data.table.tournament}
\alias{as.data.table.tournament}
\title{Convert a \code{tournament} object into a \code{data.table}}
\usage{
\method{as.data.table}{tournament}(x)
}
\arguments{
\item{x}{[object of S3 class \code{"tournament"}]\cr
the return value of \code{\link{tournament}}}
}
\value{
\code{data.table} with all tournament results
}
\description{
Convert a \code{tournament} object into a \code{data.table}
}
\seealso{
\code{\link[data.table]{data.table}}, \code{\link{tournament}}
}
