% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawPerm.r
\name{drawPerm}
\alias{drawPerm}
\title{Calculation of permutations of all possible draws}
\usage{
drawPerm(n, sample = TRUE, sample.size = 10)
}
\arguments{
\item{n}{integer(1), number of teams}

\item{sample}{logical(1), TRUE if only a sample of size sample.size
of all possible draws is returned (with replacement of draws)}

\item{sample.size}{integer(1), sample size}
}
\value{
(n x m)-dim. matrix where m=factorial(n) if sample=FALSE or
  m=sample.size if sample=TRUE.
}
\description{
Calculation of permutations of all possible draws
}
\note{
Stay away from simulating all possible draws if there are more
  than 8 teams competing in your tournament. Use sample=TRUE.
}
