% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expGoals.r
\name{expGoals}
\alias{expGoals}
\title{Calculation of expected goals}
\usage{
expGoals(x1, x2)
}
\arguments{
\item{x1}{[\code{integer(5)}]\cr
named team vector of team 1}

\item{x2}{[\code{integer(5)}]\cr
named team vector of team 2}
}
\value{
\code{numeric(1)}, expected goals of team 1 vs. team 2
}
\description{
Calculates the expected goals of \code{x1} vs \code{x2} where
  \code{ x1 = [t, a, v, m, s] } and
  \code{ x2 = [t, a, v, m, s] }.
}
\examples{
x1 <- c(t = 0,  a = 10, v = 16, m = 16, s = 48)
x2 <- c(t = 10, a = 10, v = 36, m = 12, s = 12)
expGoals(x1, x2)
expGoals(x2, x1)

}
