% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{gameRep}
\alias{gameRep}
\title{Internal helper function that calls game() repsGame-times}
\usage{
gameRep(x1, x2, etps = TRUE, method = "sample", repsGame = 4)
}
\arguments{
\item{x1}{integer(5), named team vector of team 1}

\item{x2}{integer(5), named team vector of team 2}

\item{etps}{logical(1), play etra time and penalty shoot-out? Default is TRUE.}

\item{method}{character(1), either "expected" or "sample" (default).}

\item{repsGame}{integer(1), number of replications}
}
\value{
integer(2) as in game()
}
\description{
Internal helper function that calls game() repsGame-times
}
