% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.r
\name{league}
\alias{league}
\title{Play league.}
\usage{
league(lineups, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{(N x 5)-dim. matrix of lineups where N=2,4,8,16, ...
Column names are t,a,v,m,s.
Row names are the teams.}

\item{method}{character(1), either "expected" or "sample" (default).
More details in game().}

\item{repsGame}{integer(1), number of repetitions of each game (if method="sample").
Default is 1 (i.e. no repetitions).}
}
\value{
list of class "league" that contains:
  * points, a data.table with the points reached by each team
  * results, a data.table with the full-time results
  * resultsFull, a data.table with all full-time results (i.e. repetitions of the very same game
    are included if repsGame > 1).
}
\description{
Play league.
}
\note{
* No home or away games are supported.
* If you use method="sample" and repsGame > 1, league$results will contain non-integer scores and points.
}
\examples{
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups

set.seed(2020)
(res <- league(lineups))
str(res)
res$results
league(lineups, method = "sample")                 # different from 1st run
league(lineups, method = "expected")               # always the same
league(lineups, method = "sample", repsGame = 20)  # 20x repetition of each game => more exact results
league(lineups, method = "sample", repsGame = 20)  # close to the previous league simulation

}
