% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{mergeAnalysis}
\alias{mergeAnalysis}
\title{Merge a list of \code{simResultAnalysis} objects}
\usage{
mergeAnalysis(x)
}
\arguments{
\item{x}{[\code{list}]\cr
list of \code{simResultAnalysis} objects}
}
\value{
\code{list} with results of class \code{"simResultAnalysis"}
}
\description{
Merge a list of \code{simResultAnalysis} objects
}
\seealso{
\code{\link{analysis}}
}
