% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{plot.simResultAnalysis}
\alias{plot.simResultAnalysis}
\title{Visualize the analysis results of a simulation}
\usage{
\method{plot}{simResultAnalysis}(x, type = "distrib", ...)
}
\arguments{
\item{x}{object of class \code{simResultAnalysis}}

\item{type}{[\code{character(1)}]\cr
plot type with
\code{type="distrib"} for a plot of the discrete distribution of the simulated
tournament performance of each team or
\code{type="scatter"} for a scatter plot with a comparison of the true result
and the tournament simulation results}

\item{...}{further arguments that are passed to the \code{plot} functions}
}
\value{
\code{invisible(NULL)}
}
\description{
Visualize the analysis results of a simulation
}
\seealso{
\code{\link{analysis}}
}
