% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{print.game}
\alias{print.game}
\title{Print method for game objects}
\usage{
\method{print}{game}(x, extratime = "n.V.", penalty = "i.E.", teamnames = NULL)
}
\arguments{
\item{x}{object of class "game"}

\item{extratime}{character(1), what to add for extra time result}

\item{penalty}{character(1), what to add for penalty shoot-out result}

\item{teamnames}{NULL or character(2), team names if known}
}
\value{
invisible(character(1))
}
\description{
Print method for game objects
}
