% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league.r
\name{print.league}
\alias{print.league}
\title{\code{print} method for \code{league} objects}
\usage{
\method{print}{league}(x)
}
\arguments{
\item{x}{[object of S3 class \code{"league"}]\cr
the return value of \code{\link{league}}}
}
\value{
\code{invisible(x$points)} ordered by \code{points}
}
\description{
\code{print} method for \code{league} objects
}
\seealso{
\code{\link{league}}
}
