% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{print.simResultAnalysis}
\alias{print.simResultAnalysis}
\title{\code{print} method for \code{simResultAnalysis} objects}
\usage{
\method{print}{simResultAnalysis}(x, ...)
}
\arguments{
\item{x}{[object of S3 class \code{"simResultAnalysis"}]\cr
the return value of either \code{\link{analysis}} or \code{\link{mergeAnalysis}}}

\item{...}{further arguments passed to \code{\link[base]{print.default}}}
}
\value{
\code{invisible(x)}
}
\description{
This print method basically just calls \code{\link[base]{print.default}}
but omits to print the \code{x} attribute (the input of class \code{"simResult"}).
}
