% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{print.tournament}
\alias{print.tournament}
\title{\code{print} method for \code{tournament} objects}
\usage{
\method{print}{tournament}(x)
}
\arguments{
\item{x}{[object of S3 class \code{"tournament"}]\cr
the return value of \code{\link{tournament}}}
}
\value{
\code{invisible(x)}
}
\description{
\code{print} method for \code{tournament} objects
}
