% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expGoals.r
\name{round2}
\alias{round2}
\title{"Mercantile" rounding of numbers}
\usage{
round2(x)
}
\arguments{
\item{x}{[\code{numeric(n)}]\cr
\code{n}-dim. numeric vector}
}
\value{
\code{floor(0.5 + x)}
}
\description{
\code{\link[base]{round}} can round off a 5, i.e. \code{round(0.5)} is \code{0}.
The BFB rules require "mercantile" rounding, i.e. \code{round(0.5)} should be \code{1}.
}
\examples{
z <- c(0.5, 3/2, 5/2, 1.5, 2.5, 1e3 + 0.5, -1.5, -2.5)
data.frame(z, round(z), round2(z))

}
\references{
\href{https://stackoverflow.com/questions/12688717/round-up-from-5}{Stackoverflow thread} and
\href{https://en.wikipedia.org/wiki/IEEE_754#Directed_roundings}{Wikipedia about IEEE 754}.
}
