% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLeague.r
\name{simLeague}
\alias{simLeague}
\title{Simulation of a league}
\usage{
simLeague(lineups, reps = 1000, ncores = 8, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{[\code{matrix}]\cr
(\code{N} x 5)-dim. matrix of lineups.
Column names are t,a,v,m,s.
Row names are the teams.}

\item{reps}{[\code{integer(1)}]\cr
number of replications of a league (if \code{method="sample"})}

\item{ncores}{[\code{integer(1)}]\cr
number of processor cores for parallelization}

\item{method}{[\code{character(1)}]\cr
either \code{"expected"} or \code{"sample"} (default).
More details in \code{\link{game}}.}

\item{repsGame}{[\code{integer(1)}]\cr
number of repetitions of each game (if \code{method="sample"}).
Default is 1 (i.e. no repetitions)}
}
\value{
list of S3 class \code{"simResult"} with:
\describe{
  \item{\code{results}}{  [\code{data.table}]\cr results of every game }
  \item{\code{summary}}{  [\code{data.table}]\cr points won by the teams }
  \item{\code{lineups}}{  [\code{matrix}]\cr the \code{lineups} input parameter matrix }
}
}
\description{
Simulation of a league
}
\examples{
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups

set.seed(2020)
(res <- simLeague(lineups, reps = 100))
str(res)

# For less variability in the results, the whole league can be played more often
# (increase reps) or each game can be replicated repsGame-times which results in average
# game results with an average score and average points.
(res <- simLeague(lineups, reps = 10, method = "sample", repsGame = 20))
(res <- simLeague(lineups, reps = 100, method = "sample", repsGame = 20))
# Or, method="expected" can be used:
(res <- simLeague(lineups, method = "expected"))
res$results        # score not necessarily integer-valued for method="expected".
str(res)           # only 1 leagueID, i.e. one run. No sampling of results.

}
\seealso{
\code{\link{league}}, \code{\link{simAmateurs}}
}
