% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTournament.r
\name{simTournament}
\alias{simTournament}
\title{Simulation of a tournament with fixed lineups}
\usage{
simTournament(lineups, reps = 100, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{(N x 5)-dim. matrix of lineups where N=2,4,8,16, ...
Column names are t,a,v,m,s.
Row names are the teams.}

\item{reps}{integer(1), number of replications of single tournament}

\item{method}{character(1), either "expected" or "sample" (default)
More details in game().}

\item{repsGame}{integer(1), number of repetitions of each game (if method="sample").
Default is 1 (i.e. no repetitions). repsGame > 1 is NOT recommended for tournaments.}
}
\value{
list of class "simResult" with:
  * results: data.table with the results of every game
  * summary: data.table with points won by the teams
  * lineups: the lineups input parameter matrix
  * resultsTournament: list of created tournament objects
}
\description{
Simulation of a tournament with fixed lineups
}
\examples{
# define some lineups:
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups

set.seed(2020)
(res <- simTournament(lineups, reps = 100))
res$results                # 300 = reps * (3 games per tournament with 4 teams) games alltogether
res$resultsTournament[[1]] # the 1st of reps tournaments
names(res)
str(res[1:3])
str(head(res$resultsTournament, 2))

(res <- simTournament(lineups, method = "expected"))  # no sampling of results

}
