% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{tournament}
\alias{tournament}
\title{Play tournament.}
\usage{
tournament(lineups, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{(N x 5)-dim. matrix of lineups where N=2,4,8,16, ...
Column names are t,a,v,m,s.
Row names are the teams.}

\item{method}{character(1), either "expected" or "sample" (default).
More details in game().}

\item{repsGame}{integer(1), number of repetitions of each game (if method="sample").
Default is 1 (i.e. no repetitions). repsGame > 1 is NOT recommended for tournaments.}
}
\value{
list with results of class "tournament"
}
\description{
Play tournament.
}
\note{
* The row-wise order of lineups determines the tournament:
  stage 1:
    row1 vs. row2
    row3 vs. row4
    ...
  stage 2:
    winner1 vs. winner2
    winner3 vs. winner4
    ...
  As many stages follow as necessary to determine an overall
  winner in a final stage, the final.
}
\examples{
# define some lineups:
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups
lineups2 <- rbind(lineups, lineups)
rownames(lineups2) <- paste0(rownames(lineups), c(rep("", 4), rep(" II", 4)))
lineups2

set.seed(2020)
res <- tournament(lineups)
str(res)
Points.tournament(res)
as.data.table(res)
res

tournament(lineups)
tournament(lineups)                       # different from 1st two runs
tournament(lineups, method = "expected")  # always the same

tournament(lineups2)

# For tournaments, it's really not recommended to use repsGame > 1:
set.seed(2020)
tournament(lineups)
tournament(lineups)
set.seed(2020)
(res <- tournament(lineups, repsGame = 5))
as.data.table(res)
# => Here, we we actually have decisions without extra time.

}
