% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{analysis}
\alias{analysis}
\title{Analysis of a simulation result of a tournament and league}
\usage{
analysis(x, truth)
}
\arguments{
\item{x}{[object of S3 class \code{"simResult"}]\cr
list with \code{tournament} and \code{league} entries, each of class \code{simResult}}

\item{truth}{[\code{data.frame}]\cr
with true results with columns \code{team}
and \code{tournamentRoundResult} (points won in the tournament, 1 point per game that was won)}
}
\value{
\code{list} with results of class \code{simResultAnalysis}
}
\description{
Analyse a simulation result returned by \code{\link{simAmateurs}}.
}
\examples{
5

}
\seealso{
Call \code{\link{simAmateurs}} to create an object of class \code{"simResult"}.
Use \code{\link{plot.simResultAnalysis}} for a visualization of the analysis result and
\code{\link{mergeAnalysis}} to merge two or more analysis results.
\code{\link{print.simResultAnalysis}} basically just calls \code{\link[base]{print.default}}
but omits to print the \code{x} attribute (the input of class \code{"simResult"}).
}
