% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{as.data.table.tournament}
\alias{as.data.table.tournament}
\title{Convert a tournament object into a data.table}
\usage{
\method{as.data.table}{tournament}(x)
}
\arguments{
\item{x}{list of class "tournament"}
}
\value{
invisible(x)
}
\description{
Convert a tournament object into a data.table
}
