% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expGoals.r
\name{expGoals}
\alias{expGoals}
\title{Calculates the expected goals of x1 vs x2 where
  x1 = [t, a, v, m, s]'
  x2 = [t, a, v, m, s]'}
\usage{
expGoals(x1, x2)
}
\arguments{
\item{x1}{integer(5), named team vector of team 1}

\item{x2}{integer(5), named team vector of team 2}
}
\value{
numeric(1), expected goals of team 1 vs. team 2
}
\description{
Calculates the expected goals of x1 vs x2 where
  x1 = [t, a, v, m, s]'
  x2 = [t, a, v, m, s]'
}
\examples{
x1 <- c(t = 0,  a = 10, v = 16, m = 16, s = 48)
x2 <- c(t = 10, a = 10, v = 36, m = 12, s = 12)
expGoals(x1, x2)
expGoals(x2, x1)

}
