% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{points.tournament}
\alias{points.tournament}
\title{Award points to teams participating in a tournament}
\usage{
\method{points}{tournament}(x)
}
\arguments{
\item{x}{[object of S3 class \code{"tournament"}]\cr
the return value of \code{\link{tournament}}}
}
\value{
\code{integer(}number of teams\code{)}, vector of points awarded to each team.
  1 point is awarded for each win.
}
\description{
Award points to teams participating in a tournament
}
