% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{points.tournament}
\alias{points.tournament}
\title{Award points to teams participating in a tournament}
\usage{
\method{points}{tournament}(x)
}
\arguments{
\item{x}{list of class "tournament"}
}
\value{
integer(1), vector of points awarded to each team.
  1 point is awarded for each win.
}
\description{
Award points to teams participating in a tournament
}
