% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.r
\name{print.game}
\alias{print.game}
\title{\code{print} method for \code{game} objects}
\usage{
\method{print}{game}(x, extratime = "n.V.", penalty = "i.E.", teamnames = NULL)
}
\arguments{
\item{x}{[object of S3 class \code{"game"}]\cr
a game result returned by \code{\link{game}}}

\item{extratime}{[\code{character(1)}]\cr
what to append for the extra time result}

\item{penalty}{[\code{character(1)}]\cr
what to append for the penalty shoot-out result}

\item{teamnames}{[\code{NULL} or \code{character(2)}]\cr
team names (if known)}
}
\value{
\code{invisible(character(1))},
  the nicely formatted game result
}
\description{
\code{print} method for \code{game} objects
}
\seealso{
\code{\link{game}} for a single game and \code{\link{gameRep}} for a game repeated multiple times
}
