% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simResult.r
\name{print.simResult}
\alias{print.simResult}
\title{Print method for objects of class "simResult"}
\usage{
\method{print}{simResult}(x, printNum = TRUE)
}
\arguments{
\item{x}{object of class "simResult"}

\item{printNum}{logical(1), print (and calculate) the numbers (cf. return)?}
}
\value{
invisible(list) with:
  * summary: results summary and lineups (if available)
  * nums:    integer(6) with the number of teams, tournaments, leagues, games,
             games with extra-time and games with penalty shoot-outs
}
\description{
Print method for objects of class "simResult"
}
