% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{print.tournament}
\alias{print.tournament}
\title{Print method for tournament objects}
\usage{
\method{print}{tournament}(x)
}
\arguments{
\item{x}{list of class "tournament"}
}
\value{
invisible(x)
}
\description{
Print method for tournament objects
}
