% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLineups.r
\name{readLineups}
\alias{readLineups}
\title{Read data from ZAT .ods files}
\usage{
readLineups(path = "data", nmr = "NMR-Amateure.ods")
}
\arguments{
\item{path}{character(1), path of folder with lineup files}

\item{nmr}{character(1), name of NMR file in the same folder}
}
\value{
correctly named lineup matrix that can be used as input for tournament()
}
\description{
Read data from ZAT .ods files
}
