% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTournament.r
\name{simTournament}
\alias{simTournament}
\title{Simulation of a tournament with a fixed \code{lineups} matrix}
\usage{
simTournament(lineups, reps = 100, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{[\code{matrix}]\cr
(\code{N} x 5)-dim. matrix of lineups where \code{N}=2,4,8,16, ...
Column names are t,a,v,m,s.
Row names are the teams.}

\item{reps}{[\code{integer(1)}]\cr
number of replications of the single tournament}

\item{method}{[\code{character(1)}]\cr
either \code{"expected"} or \code{"sample"} (default).
More details in \code{\link{game}}.}

\item{repsGame}{[\code{integer(1)}]\cr
number of repetitions of each game (if \code{method="sample"}).
Default is 1 (i.e. no repetitions). \code{repsGame} > 1 is NOT recommended for tournaments.}
}
\value{
list of S3 class \code{"simResult"} with:
\describe{
  \item{\code{results}}{  [\code{data.table}]\cr results of every game }
  \item{\code{summary}}{  [\code{data.table}]\cr points won by the teams }
  \item{\code{lineups}}{  [\code{matrix}]\cr the \code{lineups} input parameter matrix }
  \item{\code{resultsTournament}}{  [\code{list}]\cr all created \code{\link{tournament}} objects }
}
}
\description{
If \code{method="sample"}, the better team does not necessarily win
every game of a stage of a tournament. That is why \code{simTournament}
calls \code{\link{tournament}} \code{reps}-times and returns
an easy-to-read summary.
}
\examples{
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups

set.seed(2020)
(res <- simTournament(lineups, reps = 100))
res$results                # 300 = reps * (3 games per tournament with 4 teams) games alltogether
res$resultsTournament[[1]] # the 1st of reps tournaments
names(res)
str(res[1:3])
str(head(res$resultsTournament, 2))

(res <- simTournament(lineups, method = "expected"))  # no sampling of results

}
\seealso{
\code{\link{tournament}}, \code{\link{simTournamentPerm}}, \code{\link{simAmateurs}}
}
