% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.r
\name{tournament}
\alias{tournament}
\title{Play tournament}
\usage{
tournament(lineups, method = "sample", repsGame = 1)
}
\arguments{
\item{lineups}{[\code{matrix}]\cr
(\code{N} x 5)-dim. matrix of lineups where \code{N}=2,4,8,16, ...
Column names are t,a,v,m,s.
Row names are the teams.}

\item{method}{[\code{character(1)}]\cr
either \code{"expected"} or \code{"sample"} (default).
More details in \code{\link{game}}.}

\item{repsGame}{[\code{integer(1)}]\cr
number of repetitions of each game (if \code{method="sample"}).
Default is 1 (i.e. no repetitions). \code{repsGame} > 1 is NOT recommended for tournaments.}
}
\value{
\code{list} with results of class \code{"tournament"}
}
\description{
Play a complete tournament.
The ordering of the rows of \code{lineups} determines the tournament:
\itemize{
  \item stage 1:\cr
    row1 vs. row2 (determine winner1)\cr
    row3 vs. row4 (determine winner2)\cr
    ...
  \item stage 2:\cr
    winner1 vs. winner2\cr
    winner3 vs. winner4\cr
    ...
}
As many stages follow as necessary to determine an overall winner in a final stage, the final.
}
\examples{
# define some lineups:
lineups <- cbind(t = c(4,0,0,4), a = c(0,4,4,0), v = c(5,5,8,8), m = c(8,5,5,5), s = c(5,8,5,5))
rownames(lineups) <- c("Nob", "FdS", "USV", "Marco")
lineups
lineups2 <- rbind(lineups, lineups)
rownames(lineups2) <- paste0(rownames(lineups), c(rep("", 4), rep(" II", 4)))
lineups2

set.seed(2020)
res <- tournament(lineups)
str(res)
points(res)
as.data.table(res)
res

tournament(lineups)
tournament(lineups)                       # different from 1st two runs
tournament(lineups, method = "expected")  # always the same

tournament(lineups2)

# For tournaments, it's really not recommended to use repsGame > 1:
set.seed(2020)
tournament(lineups)
tournament(lineups)
set.seed(2020)
(res <- tournament(lineups, repsGame = 5))
as.data.table(res)
# => Here, we we actually have decisions without extra time.

}
\seealso{
\code{\link{print.tournament}}, \code{\link{points.tournament}}, \code{\link{as.data.table.tournament}},
  \code{\link{simTournament}}, \code{\link{simTournamentPerm}}, \code{\link{simAmateurs}}
}
